Rule execution: Use geometric progression for the retry delay
The naive approach to retries is not satisfactory. If there is an outage
somewhere in the request path (RESTBase, MW, etc), retrying three times
0.5 seconds apart will not help, as these are usually situations where
human intervention is needed.

We now use geometric progression to allow sufficient time to pass by so
that retries have a chance of succeeding. For that, we introduce a third
parameter - `retry_factor` - which is the basis of the progression, and
its default value is set to 6. That gives us the following delays if all
of the defaults are used: 0.5s, 3s, 18s, 1m48s, 10m48s and 1h4m48s