Rerender mobile-sectionss and summary on MW purge and null edit.
Sometimes due to software bugs MCS output could be wrong.
If it's cached, the editors will naturally try to purge it with
MW purge or a null edit. But since HTML might not have changed,
the optimization in RESTBase that doesn't issue purges if HTML's
unchanged will not be issued. This config change fixes the situation.

Bug: T182953