Commit offsets for non=matched messages from time to time.
For some rules like null edits in change-prop the rule match
is very unprobably, because they reuse a common topic with
many other rules that fire way more frequently.

If we only commit messages when they match, that can create a large
false backlog - after a restart a lot of messages can be reread and
continuosly rejected.

To avoid it, commit offsets of the rejected messages from time to time.